REQUIREMENT-AI.v01/requirements.txt
langchain>=0.2.0
openai>=1.30.0
faiss-cpu>=1.7.4
sentence-transformers>=2.7.0
transformers>=4.41.0
spacy>=3.7.4
scikit-learn>=1.5.0
scipy>=1.11.0
xgboost>=2.0.0
pymupdf>=1.24.0
pytesseract>=0.3.10
pdf2image>=1.17.0
Pillow>=10.3.0
numpy>=1.26.4
pandas>=2.2.2
torch>=2.2.0
xgboost>=2.0.0
pymupdf>=1.24.0
pytesseract>=0.3.10
pdf2image>=1.17.0
Pillow>=10.3.0
numpy>=1.26.4
pandas>=2.2.2
python-docx>=1.1.2
reportlab>=4.1.0
openpyxl>=3.1.2
rich>=13.7.1
typer>=0.12.3
uvicorn>=0.30.0
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0,<3
pydantic-core>=2.18.0,<3
streamlit>=1.36.0
python-dotenv>=1.0.1
pyyaml>=6.0.1
tiktoken>=0.7.0

